integrated with exception module and implemented interceptor to intercept all calls to the controllers and perform logging and validations

+++ expesnses-tracker-backend/src/main/java/name/expenses/features/expesnse/dao/ExpenseDAO.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface ExpenseDAO {
     Expense createExpense(Expense expense);
     Expense getExpense(String refNo);
-    Page<Expense> findAll(int pageNumber, int pageSize, String sortBy, SortDirection sortDirection);
+    Page<Expense> findAll(Long pageNumber, Long pageSize, String sortBy, SortDirection sortDirection);
     List<Expense> getAllExpenses();
     Expense updateExpense(Expense expense);
     String deleteExpense(String refNo);
