+++ expesnses-tracker-backend/src/main/java/name/expenses/features/expesnse/service/ExpenseService.java
@@ -3,14 +3,10 @@ package name.expenses.features.expesnse.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
 import jakarta.ejb.Local;
-import jakarta.ws.rs.core.Response;
 import name.expenses.features.expesnse.dtos.request.ExpenseReqDto;
 import name.expenses.features.expesnse.dtos.request.ExpenseUpdateDto;
-import name.expenses.features.expesnse.models.Expense;
 import name.expenses.globals.SortDirection;
 import name.expenses.globals.responses.ResponseDto;
-
-import java.util.List;
 
 
 @Local
@@ -22,6 +18,6 @@ public interface ExpenseService {
 
     ResponseDto deleteExpense(String refNo) throws JsonProcessingException;
 
-    ResponseDto getAllEntities(int pageNumber, int pageSize, String sortBy, SortDirection sortDirection);
+    ResponseDto getAllEntities(Long pageNumber, Long pageSize, String sortBy, SortDirection sortDirection);
 
 }
